requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
google-generative-ai
PyPDF2
python-dotenv
faiss-cpu
